pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "indexara"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "anthropic>=0.25",
    "mutagen>=1.47",
    "pdfminer.six>=20221105",
    "python-docx>=1.1",
    "pathspec>=0.12",
    "rich>=13.7",
    "typer>=0.12",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]

[project.scripts]
indexara = "indexara.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["indexara*"]
